Separated out implementation examples
